+++ semcore/notice-bubble/__tests__/notice-bubble.browser-test.tsx
@@ -143,21 +143,17 @@ test.describe('Functional', () => {
     await test.step('Verify notice shown when click on th trigger', async () => {
       await buttonTrigger.click();
       await locators.closeButton(page).waitFor({ state: 'visible' });
-      await locators.closeHint(page).waitFor({ state: 'visible' });
-      await expect(locators.closeHint(page)).toBeVisible(); // bug UIK-3926
+      await expect(locators.closeHint(page)).toBeHidden();
     });
 
     await test.step('Verify secong notice shown when click on th trigger', async () => {
       await buttonTrigger.click();
       await locators.closeButton(page).nth(1).waitFor({ state: 'visible' });
-      await locators.closeHint(page).waitFor({ state: 'visible' });
-      await expect(locators.closeHint(page)).toBeVisible(); // bug UIK-3926
+      await expect(locators.closeHint(page)).toBeHidden();
       await expect(locators.closeButton(page)).toHaveCount(2);
     });
 
     await test.step('Verify focus returns to the trigger be Escape', async () => {
-      await page.keyboard.press('Escape');
-      await locators.closeHint(page).waitFor({ state: 'hidden' });
       await page.keyboard.press('Escape');
       await locators.closeButton(page).nth(1).waitFor({ state: 'hidden' });
       await expect(buttonTrigger).toBeFocused();
@@ -357,7 +353,7 @@ test.describe('Visual', () => {
     await buttonTrigger.click();
     await buttonTrigger.click();
     await locators.closeButton(page).nth(1).waitFor({ state: 'visible' });
-    await locators.closeHint(page).waitFor({ state: 'visible' });
+    await expect(locators.closeHint(page)).toBeHidden();
     await expect(page).toHaveScreenshot();
   });