+++ semcore/side-panel/src/SidePanel.jsx
@@ -214,6 +214,7 @@ function Close({ styles, children: hasChildren, Children, getI18nText }) {
       theme='muted'
       size='l'
       addonLeft={hasChildren ? undefined : CloseIcon}
+      data-hide-focus-hover-popper='true'
     >
       {hasChildren ? <Children /> : undefined}
     </SClose>,